Fixing usage of Tpetra templated code and ETI (VRI Kanban #2852, DTK #13)
As explained in VRI Kanban #2852 and DTK #13 (GitHub issue), DTK was
incorrectly including the Tpetra_X_decl.hpp and Tpetra_X_def.hpp versions of
the Tpetra files Tpetra_X.hpp.  With this commit, DTK is now only including
the correct Tpetra_X.hpp versions of the header files and now correctly relies
on ETI in Tpetra to generate the needed instantaitions.

Also, DTK was incorrectly using the same types for LocalOrdnal and
GlobalOrdinal.  No-one ever needs more than an int for a LocalOrdinal (but
might need less).  Therefore, I have explicitly changes the LocalOrdinal to
int.

Currently, DTK tests are requesting a lot of non-standard types for
GlobalOrdinal and Scalar.  I had to add, in DataTransferKit/CMakeLists.txt, a
bunch of these instantations.  This is way overkill but it makes DTK compile
again.

A few changes had to be made to Tpetra to support the 'unsigned int' types
that will need to be pushed before these changes are pushed.