Changes to the mesh-free interpolation operators:
Operator A has been broken up into a composite operator
for efficiency. There is the polyonomial part and the basis
part with the apply operation now handled through an
implementation of the Tpetra::Operator interface in the
same way as operator C.

Both operators A and C have also been modified such that
the CrsMatrices are created with a static profile as the
number of entries per row is always known a priori. This
resulted in a performance improvement for the assembly phase.

It was discovered that the unscalable component of the
SplineInterpolator was the call to fillComplete() for the basis
part of operator A. This is likely due to the fact that
this matrix is non-square.