Removing onSurface function from Entity API
This function call didn't make a lot of sense when doing
the first mesh implementation. Technically, the surface of
a mesh is just a collection of boundaries. Furthermore, it
is not entirely clear what this collection is for any given
mesh. Therefore, we can simply ask users to assemble the surface
of interest from a list of boundary ids and construct a
selection predicate from that list.