Adding tolerance to coarse global search.
The coarse global search point inclusion tolerance adds 'fuzzy'
boundaries to the bounding boxes used in the preliminary determination
of domain destination processes for range points. This tolerance
is set as absolute using the 'Point Inclusion Tolerance' parameter
in the ParameterList. A relative version will be computed for
each bounding box in the coarse global search.