Adding option to keep range data for missed points instead of zero to consistent interpolation op
This option lets users keep the part of the range vector that has
missing points when enabled. This automatically adds tracking of
missed range entities as well as this is needed for making the vector.
To enable this in a general way the map operator now always pulls
data from the range vector application before applying the implementation
operator. We will want to review the performance of this later.