FieldMultiVector: speed up pull and push Data
ArrayRCP is inherently slow as it requires indirect referencing twice.
ArrayRCP<ArrayRCP> is even slower. I got rid of one of them. A better
version would be to start using Tpetra::MultiVector::getLocalView() to
get direct access to Kokkos.

Another issue in the code was the strided access to the support ids
vector. I rearranged the loops so we get a better access pattern.