Integral test: examining L2 integration approach
This addresses #90.

L2 projection framework does not use Jacobian when computing mass
matrix. Instead, it relies on the following code:

        // Fill the mass matrix.
        range_cardinality = range_support_ids.size();
        for ( int ni = 0; ni < range_cardinality; ++ni )
        {
            mm_values.assign( range_cardinality, 0.0 );
            for ( int p = 0; p < num_ip; ++p )
            {
                temp =
                    range_entity_measure * int_weights[p] * shape_evals[p][ni];
                for ( int nj = 0; nj < range_cardinality; ++nj )
                {
                    mm_values[nj] += temp * shape_evals[p][nj];
                }
            }
            mass_matrix->insertGlobalValues( range_support_ids[ni],
                                             range_support_ids(),
                                             mm_values() );
        }

So, instead of Jacobians it uses range_entity_measure.

Our assumption is that it is wrong to use this approach when the transformation
from a reference to physical is exactly nonlinear. This test is a sipmle example
comparing two approaches: the one in L2 projection and the one using Intrepid
integration. For a construct function f(x) = 1, the exact integral is known and
is the volume of the domain. Thus, we can calculate the quality of integration.

As expected, for non-zero perturbation the L2 way loses its accuracy.