ReferenceHexMesh: added a perturbantion parameter
This allows creation of Hex<8> mesh with nonlinear transformations.

The code creates the original Cartesian mesh and then randomly perturbs
the interanl nodes by tau*h in each dimension, where tau is a
perturbation parameter, \tau \in [0,0.5), and h is the mesh step size in
that direction (the original mesh is uniform).