L2Projection test: fix the integrate() function
The test had a bug in trying to compare integrals before and after.
Specifically, it was always comparing 0 with 0, and during the
construction of a ghosted FieldMultiVector the data was not pulled from
the original field and was not pushed into the ghosted field.
Essentially, the multivector import was importing zeros to zeros.

With a fixed function I had to change the integration epsilon from 1e-11
to 1e-10 (the integral is about 3.5e3, so the relative error is around
1e-14 anyway).