BVH: fixing insufficient array size checks
1. HAVE_DTK_DBC only works with included DTK_DBC header.
2. Fixed the asserts.

Regarding 1., it feels that DTK_DBC should *always* be included in *all*
DTK files. Otherwise, mistakes like this would keep happening.

This patch does not deal with a fundamental problem, however. The real
issue is that this particular error really belongs to
program-terminating kind of error (or at least it does with the current
API, without proper propagation of errors back to users). Therefore,
using assert is a fundamental problem here as it can be hidden by
NDEBUG, which could be triggered invisibly during DTK configuration by
CMake's RELEASE or RELWITHDEBINFO build types.