Quick workaround for the hardcoded bounds
Essentially, just add runtime checks that we do not exceed that. Proper
fix is more involved and in the current context most likely require
MemoryPool or similar for proper device allocations.

The interface for multiple dispatch is also not standardized. The
current BVH properly implements a search for a single bbox, but the
support for multiple is hacked in.