Another attempt at fix
Could not make scratch(1) work, it always segfaulted somewhere with
team_size > 1. So switched to non-hierarchical interface. However, in
that case one cannot access temporary scratch, so I switched to
providing a temporary view. That view had to be 2D view to allow for a
2D submatrices E, U, V in SVD code (there is no reshape function in
Kokkos, I believe).

Somehow, this patched and glued mess of a code seems to work.